+++ packages/core/src/utilities/table.ts
@@ -2,6 +2,17 @@ import { base } from "@/defaults/variants/stacks";
 import type { Utilities } from "@/interfaces";
 
 export const table: Utilities = {
+	"empty-cells": {
+		prefix: "ec",
+		properties: ["empty-cells"],
+		slug: "empty-cells",
+		values: {
+			h: "hide",
+			s: "show",
+		},
+		variants: base,
+	},
+
 	"caption-side": {
 		prefix: "cs",
 		properties: ["caption-side"],